compledted tasks 1579 and 1578
