+++ tdei-ui/src/components/ResetPassword/ResetPassword.js
@@ -28,8 +28,7 @@ const ResetPassword = (props) => {
     new_password: yup
       .string()
       .required("New Password is required")
-      .matches(
-        /^(?=.*[a-z])(?=.*[A-Z])(?=.*\d)(?=.*[!"#$%&'()*+,\-./:;<=>?@[\\]^_`{|}~])(?!.*\s).{8,255}$/,
+      .matches(/^(?=.*[a-z])(?=.*\d)(?=.*[A-Z])(?=.*[!\"#$%&'()*+,\-./:;<=>?@[\]^_`{|}~])(?!.*\s).{8,255}$/,
         "Password must be minimum of 8 characters in length, requires at least one lower case, one upper case, one special character and a number."
       ),
     confirm_password: yup
