+++ tdei-ui/src/routes/Datasets/DatasetRow.js
@@ -5,6 +5,7 @@ import datasetRowIcon from "../../assets/img/dataset-row.svg";
 import { workspaceUrl } from '../../services';
 import DatasetsActions from "./DatasetsActions";
 import ClipboardCopy from "../Services/ClipBoardCopy";
+import { updatedTime } from "../../utils";
 
 const DatasetRow = ({ dataset, onAction, isReleasedList }) => {
     const { metadata, data_type, service, status, uploaded_timestamp, tdei_dataset_id, project_group } = dataset;
@@ -29,25 +30,7 @@ const DatasetRow = ({ dataset, onAction, isReleasedList }) => {
         borderLeft: `8px solid ${getStatusColor()}`
     };
 
-    const updatedTime = (time) => {
-        const dateTime = new Date(time);
-        
-        const optionsDate = {
-            year: 'numeric',
-            month: 'long',
-            day: 'numeric'
-        };
-        const optionsTime = {
-            hour: '2-digit',
-            minute: '2-digit',
-            second: '2-digit',
-            hour12: false
-        };
-        const formattedDate = dateTime.toLocaleDateString('en-US', optionsDate);
-        const formattedTime = dateTime.toLocaleTimeString('en-US', optionsTime);
-        
-        return `${formattedDate}, ${formattedTime}`;
-    };
+
     const handleDropdownSelect = (eventKey) => {
         if (eventKey === 'openInWorkspace') {
             window.open(`${workspaceUrl}workspace/create/tdei?tdeiRecordId=${tdei_dataset_id}`, '_blank')?.focus();
