+++ tdei-ui/src/routes/Register/Register.js
@@ -44,8 +44,7 @@ const Register = () => {
     phone: yup.string().matches(PHONE_REGEX, "Invalid phone number"),
     password: yup
       .string()
-      .matches(
-        /^(?=.*[a-z])(?=.*[A-Z])(?=.*\d)(?=.*[!"#$%&'()*+,\-./:;<=>?@[\\]^_`{|}~])(?!.*\s).{8,255}$/,
+      .matches(/^(?=.*[a-z])(?=.*\d)(?=.*[A-Z])(?=.*[!\"#$%&'()*+,\-./:;<=>?@[\]^_`{|}~])(?!.*\s).{8,255}$/,
         "Password must be minimum of 8 characters in length, requires at least one lower case, one upper case, one special character and a number."
       ).required("Please enter password"),
     confirm: yup
